+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.9.0 (unreleased)
 
 - Introduce default version setting to DurableTaskClient and expose to orchestrator ([#393](https://github.com/microsoft/durabletask-dotnet/pull/393))
+- Add support for local credential types in DTS libraries
 
 ## v1.8.1